+++ src/Core2D/ViewModels/Shapes/LineShape.cs
@@ -76,7 +76,7 @@ namespace Core2D.Shapes
         }
 
         /// <inheritdoc/>
-        public override void Move(ISelection selection, double dx, double dy)
+        public override void Move(ISelection selection, decimal dx, decimal dy)
         {
             if (!Start.State.Flags.HasFlag(ShapeStateFlags.Connector))
             {
